working dummy guest with bare virtualization
