+++ payload/mini-hypervisor/Cargo.toml
@@ -5,4 +5,6 @@ edition = "2021"
 
 [dependencies]
 elf = { version = "0.7.2", default-features = false }
+heapless = "0.8.0"
 riscv = "0.13.0"
+spin = "0.10.0"
