+++ payload/mini-hypervisor/memory.x
@@ -1,7 +1,7 @@
 MEMORY
 {
   FLASH (rx) : ORIGIN = 0x80100000, LENGTH = 1M
-  BOOT_RAM (rw) : ORIGIN = 0x80200000, LENGTH = 256K
+  BOOT_RAM (rw) : ORIGIN = 0x80200000, LENGTH = 10M
 }
 
 REGION_ALIAS("REGION_TEXT", FLASH);
@@ -9,6 +9,7 @@ REGION_ALIAS("REGION_RODATA", FLASH);
 REGION_ALIAS("REGION_STACK", BOOT_RAM);
 
 _b_stack_size = 0x2000;
+_stack_start = ORIGIN(BOOT_RAM) + LENGTH(BOOT_RAM);
 
 SECTIONS {
