+++ payload/test_guest/.gitignore
@@ -1,0 +1,3 @@
+*.bin
+*.elf
+*.o
